Buildings are now upgradable and the upgrade status is visible in the job queue.
Changed Colony_Buildings.php to be instantiable and mimic whichever child class it represents.

Added a UI component on the main game screen that dynamically fetches and displays current jobs.

Updated Database_Row parent class with a "exists" method which will return true or false after the object has been initialized. This tells you if there was indeed an object with the specified ID in the database. To be used for error-checking.

When ajax requests can't be completed, the alert message tells you which script the failed request was trying to access.