Changed resource data for new colonies in register.php
Added a menu to the colony-management screen for constructing new buildings.

Fixed a bug in the Colony_Building class.

Fixed a bug in all children of the Colony_Building class (added 'extra_fields' to the list of extra fields).

Added a type-0 building (HQ) to colonies of new registrants in content/register.php.

Added images and descriptions for more of the buildings.