Database: Changed structure of `job_queue` table. See new structure and comments in phpMyAdmin. Database export file included in this commit. See: lygos.sql
Changed the way jobs are processed; they now allow for multiple job products instead of only buildings.

Created a constructor in Colony_Buildings and deleted the constructors in the children of that class, so they all use the parent's constructor now.

Instead of instantiating children objects of Colony_Building by calling its constructor, created a static function called construct_child that returns the child object of the specified type.