UPDATES FOR WAR ENGINE: Added algorithms for calculating fleet battles.
Added class: Mega_Fleet to represent the idea of multiple fleets working together.
Created a dynamic battle calculator in the Mega_Fleet class named fleet_battle.

Note: A ship's evasion stat should be a number between 1 and 100 representing the percent chance that the given class of ship will evade damage.

Note: battle calculator is largely untested since TJ delayed in making the ships' classes.

Added a UI element to the map so that when you click on a sector, a context menu appears with options such as "scout, attack, occupy". These options are not yet functional.

Added method "save_ships" to class "Fleet" which does the reverse action as "get_ships" (saves ship counts to database).

Added a new field `special_orders` to the `fleet_ships` table so that specific types of ships can have special orders for example to prioritize targets. The Fleet_Ship class has been updated to include this new field as well.

MISC. UPDATES:
Updated the splash page (home page) to look nicer. It now uses the blank  layout rather than the promotional layout.

Updated the website's promotional layout to include the game name/logo and prettier login buttons.

Added a quickstart page in the content folder (mostly for the "instructions" document assignment for the LMC portion of the class).

Cleaned up some code for the map. It worked before, but now it follows better practices and is more extensible.

Added the "sector info" div to the map screen, but the actual content for that div still needs to be dynamically generated.